requirements-dev.txt
-r requirements.txt
black==21.11b1
isort==5.10.1
pre-commit==2.15.0

# mypy and PEP 561 stubs
mypy==0.910
django-stubs==1.9.0
types-requests==2.26.0
